requirements.txt
# Core dependencies
openai>=1.13.0
python-dotenv>=1.0.0

# For Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
nbformat>=5.0.0

# Utilities
requests>=2.31.0
pydantic>=2.0.0
tqdm>=4.66.0

# Optional - Data processing
# pandas>=2.1.0
# numpy>=1.24.0
# matplotlib>=3.7.0

# Optional - Development tools
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.1.0 